+++ decorator.rb
@@ -1,0 +1,13 @@
+require_relative './nameable'
+
+# rubocop:disable Style/Documentation
+class Decorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
